pyproject.toml
[tool.poetry]
name = "qa-automation"
version = "0.1.0"
description = "QA Test Automation Framework - Simplified Monorepo"
authors = ["Your Team <team@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
playwright = "^1.39.0"
pytest = "^7.4.0"
pytest-playwright = "^0.4.0"
pytest-xdist = "^3.3.0"
allure-pytest = "^2.13.0"
python-dotenv = "^1.0.0"
loguru = "^0.7.2"
Faker = "^19.6.0"
pyyaml = "^6.0.3"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
ruff = "^0.0.285"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# Tool Configurations
[tool.black]
line-length = 88
target-version = ['py39']

[tool.ruff]
select = ["E", "F", "W", "I", "N"]
line-length = 88

[tool.isort]
profile = "black"

# Pytest configuration moved to pytest.ini to avoid duplication